+++ CHANGELOG.md
@@ -12,6 +12,17 @@ and this project adheres to [Semantic Versioning].
 - config: Added `advanced.crash_reporting` flag to enable reporting crashes to Baking Bad.
 - dipdup: Save tombstones in `~/.cache/dipdup` on crash.
 
+### Fixed
+
+- config: Do not perform env variable substitution in commented out lines.
+
+### Removed
+
+- cli: `--logging-config` option is removed.
+- cli: All `run` command flags are removed. Use the `advanced` section of the config.
+- cli: `cache show` and `cache clear` commands are removed.
+- config: `http.cache` flag is removed.
+
 ## [6.0.0-rc1] - 2022-07-26
 
 ### Added